refactor : ProductAdapter create, code refactor bug fix
