+++ src/main/java/com/levelUpToast/levelUpToast/controller/ProductController.java
@@ -3,14 +3,13 @@ package com.levelUpToast.levelUpToast.controller;
 import com.levelUpToast.levelUpToast.domain.UseCase.product.ProductService;
 import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
 import com.levelUpToast.levelUpToast.domain.UseCase.vendor.InfoVendor;
-import com.levelUpToast.levelUpToast.domain.UseCase.vendor.IsExistVendor;
 import com.levelUpToast.levelUpToast.domain.bodyForm.requestForm.product.ProductDeleteRequestForm;
 import com.levelUpToast.levelUpToast.domain.bodyForm.requestForm.product.ProductRequestForm;
 import com.levelUpToast.levelUpToast.domain.bodyForm.responseForm.ResponseForm;
 import com.levelUpToast.levelUpToast.domain.data.member.Member;
 import com.levelUpToast.levelUpToast.domain.data.product.ResponseProductTable;
 import com.levelUpToast.levelUpToast.domain.UseCase.member.MemberCheck;
-import com.levelUpToast.levelUpToast.domain.UseCase.product.ProblemCheck;
+import com.levelUpToast.levelUpToast.domain.UseCase.util.product.ProblemCheck;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
