+++ src/main/java/com/levelUpToast/levelUpToast/domain/UseCase/util/adapter/AdapterIsProductSeq.java
@@ -1,0 +1,4 @@
+package com.levelUpToast.levelUpToast.domain.UseCase.util.adapter;
+
+public interface AdapterIsProductSeq {
+}
