+++ src/main/java/com/levelUpToast/levelUpToast/domain/UseCase/util/adapter/ProductAdapter.java
@@ -1,5 +1,7 @@
-package com.levelUpToast.levelUpToast.domain.UseCase.product;
+package com.levelUpToast.levelUpToast.domain.UseCase.util.adapter;
 
+import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
+import com.levelUpToast.levelUpToast.domain.data.product.DataBaseProductTable;
 import com.levelUpToast.levelUpToast.domain.data.product.ResponseProductTable;
 import com.levelUpToast.levelUpToast.domain.bodyForm.requestForm.product.ProductListResponseForm;
 
@@ -13,6 +15,9 @@ public interface ProductAdapter {
 
     List<ProductListResponseForm> changeProductList(List<ResponseProductTable> responseProductTable);
 
+    ResponseProductTable changeImgToUUID(Long seq, DataBaseProductTable product) throws LevelUpToastEx;
+
+    DataBaseProductTable changeImgToSEQ(ResponseProductTable responseProductTable) throws LevelUpToastEx;
 }
