+++ src/main/java/com/levelUpToast/levelUpToast/domain/UseCase/util/product/ProductIsProduct.java
@@ -1,0 +1,7 @@
+package com.levelUpToast.levelUpToast.domain.UseCase.util.product;
+
+import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
+
+public interface ProductIsProduct {
+    void isProductSEQ(Long originalProductVendor, Long requestSeq) throws LevelUpToastEx;
+}
