+++ src/main/java/com/levelUpToast/levelUpToast/domain/repository/productRepository/productRepositoryImpl/MemoryProductRepository.java
@@ -2,6 +2,7 @@ package com.levelUpToast.levelUpToast.domain.repository.productRepository.produc
 
 import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
 import com.levelUpToast.levelUpToast.domain.UseCase.img.adapter.ImgAdapter;
+import com.levelUpToast.levelUpToast.domain.UseCase.util.adapter.ProductAdapter;
 import com.levelUpToast.levelUpToast.domain.data.product.DataBaseProductTable;
 import com.levelUpToast.levelUpToast.domain.data.product.ResponseProductTable;
 import com.levelUpToast.levelUpToast.domain.data.product.data.productinfo.DataBaseProductInfo;
@@ -22,41 +23,16 @@ public class MemoryProductRepository implements ProductRepository {
 
     private static final Map<Long, DataBaseProductTable> productStore = new ConcurrentHashMap<>();
     private final ImgAdapter simpleImgAdapter;
+
+    private final ProductAdapter productAdapter;
     private Long productSeq = 0L;
 
     private DataBaseProductTable changeImgToSEQ(ResponseProductTable responseProductTable) throws LevelUpToastEx {
-        return new DataBaseProductTable(
-                responseProductTable.getTitle(),
-                simpleImgAdapter.extractImgSeq(responseProductTable.getInitialImgUrl()),
-                responseProductTable.getTag(),
-                responseProductTable.getFunding(),
-                responseProductTable.getLike(),
-                responseProductTable.getVendorSeq(),
-                new DataBaseProductInfo(
-                        responseProductTable.getProductInfo().getText(),
-                        simpleImgAdapter.extractImgSeq(responseProductTable.getProductInfo().getProductImgUrl())
-                ),
-                responseProductTable.getBuyOption(),
-                responseProductTable.getReview()
-        );
+        return productAdapter.changeImgToSEQ(responseProductTable);
     }
 
     private ResponseProductTable changeImgToUUID(Long seq, DataBaseProductTable product) throws LevelUpToastEx {
-        ResponseProductTable changeResponseProductTable = new ResponseProductTable(
-                product.getTitle(),
-                simpleImgAdapter.extractImgUUID(product.getInitialImgUrl()),
-                product.getTag(),
-                product.getFunding(),
-                product.getLike(),
-                product.getVendorSeq(),
-                new ResponseProductInfo(
-                        product.getProductInfo().getText(),
-                        simpleImgAdapter.extractImgUUID(product.getProductInfo().getProductImgUrl())
-                ),
-                product.getBuyOption(),
-                product.getReview());
-        changeResponseProductTable.setProductSeq(seq);
-        return changeResponseProductTable;
+        return productAdapter.changeImgToUUID(seq, product);
     }
 
     @Override
