+++ src/main/java/com/levelUpToast/levelUpToast/util/product/SimpleProductChangeImg.java
@@ -1,0 +1,19 @@
+package com.levelUpToast.levelUpToast.util.product;
+
+import com.levelUpToast.levelUpToast.config.exception.LevelUpToastEx;
+import com.levelUpToast.levelUpToast.domain.UseCase.img.adapter.ImgAdapter;
+import com.levelUpToast.levelUpToast.domain.data.product.DataBaseProductTable;
+import com.levelUpToast.levelUpToast.domain.data.product.ResponseProductTable;
+import com.levelUpToast.levelUpToast.domain.data.product.data.productinfo.DataBaseProductInfo;
+import com.levelUpToast.levelUpToast.domain.data.product.data.productinfo.ResponseProductInfo;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+@RequiredArgsConstructor
+public class SimpleProductChangeImg {
+
+
+
+}
